Moving provided metric consumers into a separate package
