--- hystrix-core/src/main/java/com/netflix/hystrix/metric/HystrixCommandConstructed.java
@@ -1,54 +1,0 @@
-/**
- * Copyright 2015 Netflix, Inc.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- * http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.netflix.hystrix.metric;
-
-import com.netflix.hystrix.HystrixCommandKey;
-import com.netflix.hystrix.HystrixThreadPoolKey;
-
-/**
- * Data class that get fed to event stream when a command is constructed.
- * Was used in an experiment to get stream-based concurrency working, but not used as of 1.5.0-RC1
- */
-/* package-private*/ class HystrixCommandConstructed extends HystrixCommandEvent {
-    public HystrixCommandConstructed(HystrixCommandKey commandKey, HystrixThreadPoolKey threadPoolKey) {
-        super(commandKey, threadPoolKey);
-    }
-
-    @Override
-    public boolean isExecutionStart() {
-        return false;
-    }
-
-    @Override
-    public boolean isExecutedInThread() {
-        return false;
-    }
-
-    @Override
-    public boolean isResponseThreadPoolRejected() {
-        return false;
-    }
-
-    @Override
-    public boolean isCommandCompletion() {
-        return false;
-    }
-
-    @Override
-    public boolean didCommandExecute() {
-        return false;
-    }
-}
