+++ hystrix-core/src/main/java/com/netflix/hystrix/metric/consumer/RollingConcurrencyStream.java
@@ -13,9 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.hystrix.metric;
+package com.netflix.hystrix.metric.consumer;
 
 import com.netflix.hystrix.HystrixCommandProperties;
+import com.netflix.hystrix.metric.HystrixCommandExecutionStarted;
+import com.netflix.hystrix.metric.HystrixEventStream;
 import rx.Observable;
 import rx.Subscription;
 import rx.functions.Func0;
