+++ hystrix-core/src/test/java/com/netflix/hystrix/metric/consumer/CumulativeCommandEventCounterStreamTest.java
@@ -1,4 +1,4 @@
-package com.netflix.hystrix.metric;
+package com.netflix.hystrix.metric.consumer;
 
 import com.netflix.hystrix.HystrixCommandGroupKey;
 import com.netflix.hystrix.HystrixCommandKey;
@@ -6,6 +6,7 @@ import com.netflix.hystrix.HystrixCommandMetrics;
 import com.netflix.hystrix.HystrixCommandProperties;
 import com.netflix.hystrix.HystrixEventType;
 import com.netflix.hystrix.HystrixRequestLog;
+import com.netflix.hystrix.metric.CommandStreamTest;
 import com.netflix.hystrix.strategy.concurrency.HystrixContextRunnable;
 import com.netflix.hystrix.strategy.concurrency.HystrixRequestContext;
 import org.junit.After;
@@ -20,11 +21,11 @@ import java.util.concurrent.TimeUnit;
 
 import static org.junit.Assert.*;
 
-public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
+public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
     HystrixRequestContext context;
-    RollingCommandEventCounterStream stream;
+    CumulativeCommandEventCounterStream stream;
 
-    static HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("RollingCommandCounter");
+    private static HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("CumulativeCommandCounter");
 
     private static Subscriber<long[]> getSubscriber(final CountDownLatch latch) {
         return new Subscriber<long[]>() {
@@ -54,13 +55,13 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
     public void tearDown() {
         context.shutdown();
         stream.unsubscribe();
-        RollingCommandEventCounterStream.reset();
+        CumulativeCommandEventCounterStream.reset();
     }
 
     @Test
     public void testEmptyStreamProducesZeros() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-A");
-        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-A");
+        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
@@ -73,21 +74,20 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
         } catch (InterruptedException ex) {
             fail("Interrupted ex");
         }
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertEquals(HystrixEventType.values().length, stream.getLatest().length);
         assertFalse(hasData(stream.getLatest()));
     }
 
     @Test
     public void testSingleSuccess() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-B");
-        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-B");
+        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
 
-        CommandStreamTest.Command cmd = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 20);
+        Command cmd = Command.from(groupKey, key, HystrixEventType.SUCCESS, 20);
 
         cmd.observe();
 
@@ -99,20 +99,19 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
         assertEquals(HystrixEventType.values().length, stream.getLatest().length);
         long[] expected = new long[HystrixEventType.values().length];
         expected[HystrixEventType.SUCCESS.ordinal()] = 1;
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testSingleFailure() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-C");
-        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-C");
+        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
 
-        CommandStreamTest.Command cmd = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
+        Command cmd = Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
 
         cmd.observe();
 
@@ -125,20 +124,19 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
         long[] expected = new long[HystrixEventType.values().length];
         expected[HystrixEventType.FAILURE.ordinal()] = 1;
         expected[HystrixEventType.FALLBACK_SUCCESS.ordinal()] = 1;
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testSingleTimeout() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-D");
-        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-D");
+        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
 
-        CommandStreamTest.Command cmd = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.TIMEOUT);
+        Command cmd = Command.from(groupKey, key, HystrixEventType.TIMEOUT);
 
         cmd.observe();
 
@@ -151,20 +149,19 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
         long[] expected = new long[HystrixEventType.values().length];
         expected[HystrixEventType.TIMEOUT.ordinal()] = 1;
         expected[HystrixEventType.FALLBACK_SUCCESS.ordinal()] = 1;
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testSingleBadRequest() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-E");
-        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-E");
+        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
 
-        CommandStreamTest.Command cmd = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.BAD_REQUEST);
+        Command cmd = Command.from(groupKey, key, HystrixEventType.BAD_REQUEST);
 
         cmd.observe();
 
@@ -177,22 +174,21 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
         long[] expected = new long[HystrixEventType.values().length];
         expected[HystrixEventType.BAD_REQUEST.ordinal()] = 1;
         expected[HystrixEventType.EXCEPTION_THROWN.ordinal()] = 1;
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testRequestFromCache() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-F");
-        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-F");
+        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
 
-        CommandStreamTest.Command cmd1 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 20);
-        CommandStreamTest.Command cmd2 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.RESPONSE_FROM_CACHE);
-        CommandStreamTest.Command cmd3 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.RESPONSE_FROM_CACHE);
+        Command cmd1 = Command.from(groupKey, key, HystrixEventType.SUCCESS, 20);
+        Command cmd2 = Command.from(groupKey, key, HystrixEventType.RESPONSE_FROM_CACHE);
+        Command cmd3 = Command.from(groupKey, key, HystrixEventType.RESPONSE_FROM_CACHE);
 
         cmd1.observe();
         cmd2.observe();
@@ -209,14 +205,13 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
         long[] expected = new long[HystrixEventType.values().length];
         expected[HystrixEventType.SUCCESS.ordinal()] = 1;
         expected[HystrixEventType.RESPONSE_FROM_CACHE.ordinal()] = 2;
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testShortCircuited() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-G");
-        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-G");
+        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
@@ -225,12 +220,12 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
         //3 failures in a row will trip circuit.  let bucket roll once then submit 2 requests.
         //should see 3 FAILUREs and 2 SHORT_CIRCUITs and then 5 FALLBACK_SUCCESSes
 
-        CommandStreamTest.Command failure1 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
-        CommandStreamTest.Command failure2 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
-        CommandStreamTest.Command failure3 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
+        Command failure1 = Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
+        Command failure2 = Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
+        Command failure3 = Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
 
-        CommandStreamTest.Command shortCircuit1 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS);
-        CommandStreamTest.Command shortCircuit2 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS);
+        Command shortCircuit1 = Command.from(groupKey, key, HystrixEventType.SUCCESS);
+        Command shortCircuit2 = Command.from(groupKey, key, HystrixEventType.SUCCESS);
 
         failure1.observe();
         failure2.observe();
@@ -259,14 +254,13 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
         expected[HystrixEventType.FAILURE.ordinal()] = 3;
         expected[HystrixEventType.SHORT_CIRCUITED.ordinal()] = 2;
         expected[HystrixEventType.FALLBACK_SUCCESS.ordinal()] = 5;
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testSemaphoreRejected() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-H");
-        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-H");
+        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
@@ -276,16 +270,16 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
         //submit 2 more requests and they should be SEMAPHORE_REJECTED
         //should see 10 SUCCESSes, 2 SEMAPHORE_REJECTED and 2 FALLBACK_SUCCESSes
 
-        List<CommandStreamTest.Command> saturators = new ArrayList<CommandStreamTest.Command>();
+        List<Command> saturators = new ArrayList<Command>();
 
         for (int i = 0; i < 10; i++) {
-            saturators.add(CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 200, HystrixCommandProperties.ExecutionIsolationStrategy.SEMAPHORE));
+            saturators.add(Command.from(groupKey, key, HystrixEventType.SUCCESS, 200, HystrixCommandProperties.ExecutionIsolationStrategy.SEMAPHORE));
         }
 
-        CommandStreamTest.Command rejected1 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 0, HystrixCommandProperties.ExecutionIsolationStrategy.SEMAPHORE);
-        CommandStreamTest.Command rejected2 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 0, HystrixCommandProperties.ExecutionIsolationStrategy.SEMAPHORE);
+        Command rejected1 = Command.from(groupKey, key, HystrixEventType.SUCCESS, 0, HystrixCommandProperties.ExecutionIsolationStrategy.SEMAPHORE);
+        Command rejected2 = Command.from(groupKey, key, HystrixEventType.SUCCESS, 0, HystrixCommandProperties.ExecutionIsolationStrategy.SEMAPHORE);
 
-        for (final CommandStreamTest.Command saturator : saturators) {
+        for (final Command saturator : saturators) {
             new Thread(new HystrixContextRunnable(new Runnable() {
                 @Override
                 public void run() {
@@ -317,14 +311,13 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
         expected[HystrixEventType.SUCCESS.ordinal()] = 10;
         expected[HystrixEventType.SEMAPHORE_REJECTED.ordinal()] = 2;
         expected[HystrixEventType.FALLBACK_SUCCESS.ordinal()] = 2;
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testThreadPoolRejected() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-I");
-        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-I");
+        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
@@ -334,16 +327,16 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
         //submit 2 more requests and they should be THREADPOOL_REJECTED
         //should see 10 SUCCESSes, 2 THREADPOOL_REJECTED and 2 FALLBACK_SUCCESSes
 
-        List<CommandStreamTest.Command> saturators = new ArrayList<CommandStreamTest.Command>();
+        List<Command> saturators = new ArrayList<Command>();
 
         for (int i = 0; i < 10; i++) {
-            saturators.add(CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 200));
+            saturators.add(Command.from(groupKey, key, HystrixEventType.SUCCESS, 200));
         }
 
-        CommandStreamTest.Command rejected1 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 0);
-        CommandStreamTest.Command rejected2 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 0);
+        Command rejected1 = Command.from(groupKey, key, HystrixEventType.SUCCESS, 0);
+        Command rejected2 = Command.from(groupKey, key, HystrixEventType.SUCCESS, 0);
 
-        for (final CommandStreamTest.Command saturator : saturators) {
+        for (final Command saturator : saturators) {
             saturator.observe();
         }
 
@@ -375,14 +368,14 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
 
     @Test
     public void testFallbackFailure() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-J");
-        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-J");
+        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
 
-        CommandStreamTest.Command cmd = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_FAILURE);
+        Command cmd = Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_FAILURE);
 
         cmd.observe();
 
@@ -396,20 +389,19 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
         expected[HystrixEventType.FAILURE.ordinal()] = 1;
         expected[HystrixEventType.FALLBACK_FAILURE.ordinal()] = 1;
         expected[HystrixEventType.EXCEPTION_THROWN.ordinal()] = 1;
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testFallbackMissing() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-K");
-        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-K");
+        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
 
-        CommandStreamTest.Command cmd = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_MISSING);
+        Command cmd = Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_MISSING);
 
         cmd.observe();
 
@@ -428,8 +420,8 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
 
     @Test
     public void testFallbackRejection() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-L");
-        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-L");
+        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
@@ -438,15 +430,15 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
         //fallback semaphore size is 5.  So let 5 commands saturate that semaphore, then
         //let 2 more commands go to fallback.  they should get rejected by the fallback-semaphore
 
-        List<CommandStreamTest.Command> fallbackSaturators = new ArrayList<CommandStreamTest.Command>();
+        List<Command> fallbackSaturators = new ArrayList<Command>();
         for (int i = 0; i < 5; i++) {
-            fallbackSaturators.add(CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_SUCCESS, 400));
+            fallbackSaturators.add(Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_SUCCESS, 400));
         }
 
-        CommandStreamTest.Command rejection1 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_SUCCESS, 0);
-        CommandStreamTest.Command rejection2 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_SUCCESS, 0);
+        Command rejection1 = Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_SUCCESS, 0);
+        Command rejection2 = Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_SUCCESS, 0);
 
-        for (CommandStreamTest.Command saturator: fallbackSaturators) {
+        for (Command saturator: fallbackSaturators) {
             saturator.observe();
         }
 
@@ -470,14 +462,13 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
         expected[HystrixEventType.FALLBACK_SUCCESS.ordinal()] = 5;
         expected[HystrixEventType.FALLBACK_REJECTION.ordinal()] = 2;
         expected[HystrixEventType.EXCEPTION_THROWN.ordinal()] = 2;
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testCollapsed() {
         HystrixCommandKey key = HystrixCommandKey.Factory.asKey("BatchCommand");
-        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
+        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
@@ -501,18 +492,18 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
     }
 
     @Test
-    public void testMultipleEventsOverTimeGetStoredAndAgeOut() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-M");
-        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
+    public void testMultipleEventsOverTimeGetStoredAndNeverAgeOut() {
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-M");
+        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
-        //by doing a take(30), we ensure that all rolling counts go back to 0
+        //by doing a take(30), we ensure that no rolling out of window takes place
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(30).subscribe(getSubscriber(latch));
 
-        CommandStreamTest.Command cmd1 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 20);
-        CommandStreamTest.Command cmd2 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 10);
+        Command cmd1 = Command.from(groupKey, key, HystrixEventType.SUCCESS, 20);
+        Command cmd2 = Command.from(groupKey, key, HystrixEventType.FAILURE, 10);
 
         cmd1.observe();
         cmd2.observe();
@@ -524,7 +515,9 @@ public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
         }
         assertEquals(HystrixEventType.values().length, stream.getLatest().length);
         long[] expected = new long[HystrixEventType.values().length];
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
+        expected[HystrixEventType.SUCCESS.ordinal()] = 1;
+        expected[HystrixEventType.FAILURE.ordinal()] = 1;
+        expected[HystrixEventType.FALLBACK_SUCCESS.ordinal()] = 1;
         assertArrayEquals(expected, stream.getLatest());
     }
 }
