+++ hystrix-core/src/test/java/com/netflix/hystrix/metric/consumer/CumulativeThreadPoolEventCounterStreamTest.java
@@ -1,11 +1,13 @@
-package com.netflix.hystrix.metric;
+package com.netflix.hystrix.metric.consumer;
 
 import com.netflix.hystrix.HystrixCommandGroupKey;
 import com.netflix.hystrix.HystrixCommandKey;
-import com.netflix.hystrix.HystrixCommandMetrics;
 import com.netflix.hystrix.HystrixCommandProperties;
 import com.netflix.hystrix.HystrixEventType;
 import com.netflix.hystrix.HystrixRequestLog;
+import com.netflix.hystrix.HystrixThreadPoolKey;
+import com.netflix.hystrix.HystrixThreadPoolMetrics;
+import com.netflix.hystrix.metric.CommandStreamTest;
 import com.netflix.hystrix.strategy.concurrency.HystrixContextRunnable;
 import com.netflix.hystrix.strategy.concurrency.HystrixRequestContext;
 import org.junit.After;
@@ -20,14 +22,12 @@ import java.util.concurrent.TimeUnit;
 
 import static org.junit.Assert.*;
 
-public class HealthCountsStreamTest extends CommandStreamTest {
+public class CumulativeThreadPoolEventCounterStreamTest extends CommandStreamTest {
     HystrixRequestContext context;
-    HealthCountsStream stream;
+    CumulativeThreadPoolEventCounterStream stream;
 
-    static HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("HealthCounts");
-
-    private static Subscriber<HystrixCommandMetrics.HealthCounts> getSubscriber(final CountDownLatch latch) {
-        return new Subscriber<HystrixCommandMetrics.HealthCounts>() {
+    private static Subscriber<long[]> getSubscriber(final CountDownLatch latch) {
+        return new Subscriber<long[]>() {
             @Override
             public void onCompleted() {
                 latch.countDown();
@@ -39,8 +39,8 @@ public class HealthCountsStreamTest extends CommandStreamTest {
             }
 
             @Override
-            public void onNext(HystrixCommandMetrics.HealthCounts healthCounts) {
-                System.out.println("OnNext @ " + System.currentTimeMillis() + " : " + healthCounts);
+            public void onNext(long[] eventCounts) {
+                System.out.println("OnNext @ " + System.currentTimeMillis() + " : " + eventCounts[0] + " : " + eventCounts[1]);
             }
         };
     }
@@ -59,8 +59,11 @@ public class HealthCountsStreamTest extends CommandStreamTest {
 
     @Test
     public void testEmptyStreamProducesZeros() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-Health-A");
-        stream = HealthCountsStream.getInstance(key, 10, 100);
+        HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("Cumulative-ThreadPool-A");
+        HystrixThreadPoolKey threadPoolKey = HystrixThreadPoolKey.Factory.asKey("Cumulative-ThreadPool-A");
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("Cumulative-Counter-A");
+        stream = CumulativeThreadPoolEventCounterStream.getInstance(threadPoolKey, 10, 100);
+        stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
@@ -72,15 +75,18 @@ public class HealthCountsStreamTest extends CommandStreamTest {
         } catch (InterruptedException ex) {
             fail("Interrupted ex");
         }
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
-        assertEquals(0L, stream.getLatest().getErrorCount());
-        assertEquals(0L, stream.getLatest().getTotalRequests());
+        assertEquals(2, stream.getLatest().length);
+        assertEquals(0, stream.getLatestCount(HystrixEventType.ThreadPool.EXECUTED));
+        assertEquals(0, stream.getLatestCount(HystrixEventType.ThreadPool.REJECTED));
     }
 
     @Test
     public void testSingleSuccess() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-Health-B");
-        stream = HealthCountsStream.getInstance(key, 10, 100);
+        HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("Cumulative-ThreadPool-B");
+        HystrixThreadPoolKey threadPoolKey = HystrixThreadPoolKey.Factory.asKey("Cumulative-ThreadPool-B");
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("Cumulative-Counter-B");
+        stream = CumulativeThreadPoolEventCounterStream.getInstance(threadPoolKey, 10, 100);
+        stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
@@ -94,15 +100,18 @@ public class HealthCountsStreamTest extends CommandStreamTest {
         } catch (InterruptedException ex) {
             fail("Interrupted ex");
         }
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
-        assertEquals(0L, stream.getLatest().getErrorCount());
-        assertEquals(1L, stream.getLatest().getTotalRequests());
+        assertEquals(2, stream.getLatest().length);
+        assertEquals(1, stream.getLatestCount(HystrixEventType.ThreadPool.EXECUTED));
+        assertEquals(0, stream.getLatestCount(HystrixEventType.ThreadPool.REJECTED));
     }
 
     @Test
     public void testSingleFailure() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-Health-C");
-        stream = HealthCountsStream.getInstance(key, 10, 100);
+        HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("Cumulative-ThreadPool-C");
+        HystrixThreadPoolKey threadPoolKey = HystrixThreadPoolKey.Factory.asKey("Cumulative-ThreadPool-C");
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("Cumulative-Counter-C");
+        stream = CumulativeThreadPoolEventCounterStream.getInstance(threadPoolKey, 10, 100);
+        stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
@@ -116,15 +125,18 @@ public class HealthCountsStreamTest extends CommandStreamTest {
         } catch (InterruptedException ex) {
             fail("Interrupted ex");
         }
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
-        assertEquals(1L, stream.getLatest().getErrorCount());
-        assertEquals(1L, stream.getLatest().getTotalRequests());
+        assertEquals(2, stream.getLatest().length);
+        assertEquals(1, stream.getLatestCount(HystrixEventType.ThreadPool.EXECUTED));
+        assertEquals(0, stream.getLatestCount(HystrixEventType.ThreadPool.REJECTED));
     }
 
     @Test
     public void testSingleTimeout() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-Health-D");
-        stream = HealthCountsStream.getInstance(key, 10, 100);
+        HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("Cumulative-ThreadPool-D");
+        HystrixThreadPoolKey threadPoolKey = HystrixThreadPoolKey.Factory.asKey("Cumulative-ThreadPool-D");
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("Cumulative-Counter-D");
+        stream = CumulativeThreadPoolEventCounterStream.getInstance(threadPoolKey, 10, 100);
+        stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
@@ -138,15 +150,18 @@ public class HealthCountsStreamTest extends CommandStreamTest {
         } catch (InterruptedException ex) {
             fail("Interrupted ex");
         }
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
-        assertEquals(1L, stream.getLatest().getErrorCount());
-        assertEquals(1L, stream.getLatest().getTotalRequests());
+        assertEquals(2, stream.getLatest().length);
+        assertEquals(1, stream.getLatestCount(HystrixEventType.ThreadPool.EXECUTED));
+        assertEquals(0, stream.getLatestCount(HystrixEventType.ThreadPool.REJECTED));
     }
 
     @Test
     public void testSingleBadRequest() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-Health-E");
-        stream = HealthCountsStream.getInstance(key, 10, 100);
+        HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("Cumulative-ThreadPool-E");
+        HystrixThreadPoolKey threadPoolKey = HystrixThreadPoolKey.Factory.asKey("Cumulative-ThreadPool-E");
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("Cumulative-Counter-E");
+        stream = CumulativeThreadPoolEventCounterStream.getInstance(threadPoolKey, 10, 100);
+        stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
@@ -160,15 +175,18 @@ public class HealthCountsStreamTest extends CommandStreamTest {
         } catch (InterruptedException ex) {
             fail("Interrupted ex");
         }
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
-        assertEquals(0L, stream.getLatest().getErrorCount());
-        assertEquals(0L, stream.getLatest().getTotalRequests());
+        assertEquals(2, stream.getLatest().length);
+        assertEquals(1, stream.getLatestCount(HystrixEventType.ThreadPool.EXECUTED));
+        assertEquals(0, stream.getLatestCount(HystrixEventType.ThreadPool.REJECTED));
     }
 
     @Test
     public void testRequestFromCache() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-Health-F");
-        stream = HealthCountsStream.getInstance(key, 10, 100);
+        HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("Cumulative-ThreadPool-F");
+        HystrixThreadPoolKey threadPoolKey = HystrixThreadPoolKey.Factory.asKey("Cumulative-ThreadPool-F");
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("Cumulative-Counter-F");
+        stream = CumulativeThreadPoolEventCounterStream.getInstance(threadPoolKey, 10, 100);
+        stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
@@ -188,20 +206,26 @@ public class HealthCountsStreamTest extends CommandStreamTest {
         }
 
         System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
-        assertEquals(0L, stream.getLatest().getErrorCount());
-        assertEquals(1L, stream.getLatest().getTotalRequests()); //responses from cache should not show up here
+
+        //RESPONSE_FROM_CACHE should not show up at all in thread pool counters - just the success
+        assertEquals(2, stream.getLatest().length);
+        assertEquals(1, stream.getLatestCount(HystrixEventType.ThreadPool.EXECUTED));
+        assertEquals(0, stream.getLatestCount(HystrixEventType.ThreadPool.REJECTED));
     }
 
     @Test
     public void testShortCircuited() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-Health-G");
-        stream = HealthCountsStream.getInstance(key, 10, 100);
+        HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("Cumulative-ThreadPool-G");
+        HystrixThreadPoolKey threadPoolKey = HystrixThreadPoolKey.Factory.asKey("Cumulative-ThreadPool-G");
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("Cumulative-Counter-G");
+        stream = CumulativeThreadPoolEventCounterStream.getInstance(threadPoolKey, 10, 100);
+        stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
 
         //3 failures in a row will trip circuit.  let bucket roll once then submit 2 requests.
-        //should see 3 FAILUREs and 2 SHORT_CIRCUITs and then 5 FALLBACK_SUCCESSes
+        //should see 3 FAILUREs and 2 SHORT_CIRCUITs and each should see a FALLBACK_SUCCESS
 
         CommandStreamTest.Command failure1 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
         CommandStreamTest.Command failure2 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
@@ -229,18 +253,23 @@ public class HealthCountsStreamTest extends CommandStreamTest {
             fail("Interrupted ex");
         }
 
+        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertTrue(shortCircuit1.isResponseShortCircuited());
         assertTrue(shortCircuit2.isResponseShortCircuited());
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
-        //should only see failures here, not SHORT-CIRCUITS
-        assertEquals(3L, stream.getLatest().getErrorCount());
-        assertEquals(3L, stream.getLatest().getTotalRequests());
+
+        //only the FAILUREs should show up in thread pool counters
+        assertEquals(2, stream.getLatest().length);
+        assertEquals(3, stream.getLatestCount(HystrixEventType.ThreadPool.EXECUTED));
+        assertEquals(0, stream.getLatestCount(HystrixEventType.ThreadPool.REJECTED));
     }
 
     @Test
     public void testSemaphoreRejected() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-Health-H");
-        stream = HealthCountsStream.getInstance(key, 10, 100);
+        HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("Cumulative-ThreadPool-H");
+        HystrixThreadPoolKey threadPoolKey = HystrixThreadPoolKey.Factory.asKey("Cumulative-ThreadPool-H");
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("Cumulative-Counter-H");
+        stream = CumulativeThreadPoolEventCounterStream.getInstance(threadPoolKey, 10, 100);
+        stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
@@ -285,14 +314,20 @@ public class HealthCountsStreamTest extends CommandStreamTest {
         System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertTrue(rejected1.isResponseSemaphoreRejected());
         assertTrue(rejected2.isResponseSemaphoreRejected());
-        assertEquals(2L, stream.getLatest().getErrorCount());
-        assertEquals(12L, stream.getLatest().getTotalRequests());
+
+        //none of these got executed on a thread-pool, so thread pool metrics should be 0
+        assertEquals(2, stream.getLatest().length);
+        assertEquals(0, stream.getLatestCount(HystrixEventType.ThreadPool.EXECUTED));
+        assertEquals(0, stream.getLatestCount(HystrixEventType.ThreadPool.REJECTED));
     }
 
     @Test
     public void testThreadPoolRejected() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-Health-I");
-        stream = HealthCountsStream.getInstance(key, 10, 100);
+        HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("Cumulative-ThreadPool-I");
+        HystrixThreadPoolKey threadPoolKey = HystrixThreadPoolKey.Factory.asKey("Cumulative-ThreadPool-I");
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("Cumulative-Counter-I");
+        stream = CumulativeThreadPoolEventCounterStream.getInstance(threadPoolKey, 10, 100);
+        stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
@@ -332,14 +367,20 @@ public class HealthCountsStreamTest extends CommandStreamTest {
         System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertTrue(rejected1.isResponseThreadPoolRejected());
         assertTrue(rejected2.isResponseThreadPoolRejected());
-        assertEquals(2L, stream.getLatest().getErrorCount());
-        assertEquals(12L, stream.getLatest().getTotalRequests());
+
+        //all 12 commands got submitted to thread pool, 10 accepted, 2 rejected is expected
+        assertEquals(2, stream.getLatest().length);
+        assertEquals(10, stream.getLatestCount(HystrixEventType.ThreadPool.EXECUTED));
+        assertEquals(2, stream.getLatestCount(HystrixEventType.ThreadPool.REJECTED));
     }
 
     @Test
     public void testFallbackFailure() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-Health-J");
-        stream = HealthCountsStream.getInstance(key, 10, 100);
+        HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("Cumulative-ThreadPool-J");
+        HystrixThreadPoolKey threadPoolKey = HystrixThreadPoolKey.Factory.asKey("Cumulative-ThreadPool-J");
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("Cumulative-Counter-J");
+        stream = CumulativeThreadPoolEventCounterStream.getInstance(threadPoolKey, 10, 100);
+        stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
@@ -353,15 +394,18 @@ public class HealthCountsStreamTest extends CommandStreamTest {
         } catch (InterruptedException ex) {
             fail("Interrupted ex");
         }
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
-        assertEquals(1L, stream.getLatest().getErrorCount());
-        assertEquals(1L, stream.getLatest().getTotalRequests());
+        assertEquals(2, stream.getLatest().length);
+        assertEquals(1, stream.getLatestCount(HystrixEventType.ThreadPool.EXECUTED));
+        assertEquals(0, stream.getLatestCount(HystrixEventType.ThreadPool.REJECTED));
     }
 
     @Test
     public void testFallbackMissing() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-Health-K");
-        stream = HealthCountsStream.getInstance(key, 10, 100);
+        HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("Cumulative-ThreadPool-K");
+        HystrixThreadPoolKey threadPoolKey = HystrixThreadPoolKey.Factory.asKey("Cumulative-ThreadPool-K");
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("Cumulative-Counter-K");
+        stream = CumulativeThreadPoolEventCounterStream.getInstance(threadPoolKey, 10, 100);
+        stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
@@ -375,14 +419,18 @@ public class HealthCountsStreamTest extends CommandStreamTest {
         } catch (InterruptedException ex) {
             fail("Interrupted ex");
         }
-        assertEquals(1L, stream.getLatest().getErrorCount());
-        assertEquals(1L, stream.getLatest().getTotalRequests());
+        assertEquals(2, stream.getLatest().length);
+        assertEquals(1, stream.getLatestCount(HystrixEventType.ThreadPool.EXECUTED));
+        assertEquals(0, stream.getLatestCount(HystrixEventType.ThreadPool.REJECTED));
     }
 
     @Test
     public void testFallbackRejection() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-Health-L");
-        stream = HealthCountsStream.getInstance(key, 10, 100);
+        HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("Cumulative-ThreadPool-L");
+        HystrixThreadPoolKey threadPoolKey = HystrixThreadPoolKey.Factory.asKey("Cumulative-ThreadPool-L");
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("Cumulative-Counter-L");
+        stream = CumulativeThreadPoolEventCounterStream.getInstance(threadPoolKey, 10, 100);
+        stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
@@ -416,17 +464,21 @@ public class HealthCountsStreamTest extends CommandStreamTest {
         } catch (InterruptedException ex) {
             fail("Interrupted ex");
         }
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
-        assertEquals(7L, stream.getLatest().getErrorCount());
-        assertEquals(7L, stream.getLatest().getTotalRequests());
+
+        //all 7 commands executed on-thread, so should be executed according to thread-pool metrics
+        assertEquals(2, stream.getLatest().length);
+        assertEquals(7, stream.getLatestCount(HystrixEventType.ThreadPool.EXECUTED));
+        assertEquals(0, stream.getLatestCount(HystrixEventType.ThreadPool.REJECTED));
     }
 
+    //in a rolling window, take(30) would age out all counters.  in the cumulative count, we expect them to remain non-zero forever
     @Test
-    public void testMultipleEventsOverTimeGetStoredAndAgeOut() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-Health-M");
-        stream = HealthCountsStream.getInstance(key, 10, 100);
-
-        //by doing a take(30), we ensure that all rolling counts go back to 0
+    public void testMultipleEventsOverTimeGetStoredAndDoNotAgeOut() {
+        HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("Cumulative-ThreadPool-M");
+        HystrixThreadPoolKey threadPoolKey = HystrixThreadPoolKey.Factory.asKey("Cumulative-ThreadPool-M");
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("Cumulative-Counter-M");
+        stream = CumulativeThreadPoolEventCounterStream.getInstance(threadPoolKey, 10, 100);
+        stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(30).subscribe(getSubscriber(latch));
@@ -442,8 +494,10 @@ public class HealthCountsStreamTest extends CommandStreamTest {
         } catch (InterruptedException ex) {
             fail("Interrupted ex");
         }
-        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
-        assertEquals(0L, stream.getLatest().getErrorCount());
-        assertEquals(0L, stream.getLatest().getTotalRequests());
+
+        //all commands should have aged out
+        assertEquals(2, stream.getLatest().length);
+        assertEquals(2, stream.getLatestCount(HystrixEventType.ThreadPool.EXECUTED));
+        assertEquals(0, stream.getLatestCount(HystrixEventType.ThreadPool.REJECTED));
     }
 }
