+++ hystrix-core/src/test/java/com/netflix/hystrix/metric/consumer/RollingCommandEventCounterStreamTest.java
@@ -1,4 +1,4 @@
-package com.netflix.hystrix.metric;
+package com.netflix.hystrix.metric.consumer;
 
 import com.netflix.hystrix.HystrixCommandGroupKey;
 import com.netflix.hystrix.HystrixCommandKey;
@@ -6,6 +6,7 @@ import com.netflix.hystrix.HystrixCommandMetrics;
 import com.netflix.hystrix.HystrixCommandProperties;
 import com.netflix.hystrix.HystrixEventType;
 import com.netflix.hystrix.HystrixRequestLog;
+import com.netflix.hystrix.metric.CommandStreamTest;
 import com.netflix.hystrix.strategy.concurrency.HystrixContextRunnable;
 import com.netflix.hystrix.strategy.concurrency.HystrixRequestContext;
 import org.junit.After;
@@ -20,11 +21,11 @@ import java.util.concurrent.TimeUnit;
 
 import static org.junit.Assert.*;
 
-public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
+public class RollingCommandEventCounterStreamTest extends CommandStreamTest {
     HystrixRequestContext context;
-    CumulativeCommandEventCounterStream stream;
+    RollingCommandEventCounterStream stream;
 
-    private static HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("CumulativeCommandCounter");
+    static HystrixCommandGroupKey groupKey = HystrixCommandGroupKey.Factory.asKey("RollingCommandCounter");
 
     private static Subscriber<long[]> getSubscriber(final CountDownLatch latch) {
         return new Subscriber<long[]>() {
@@ -54,13 +55,13 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
     public void tearDown() {
         context.shutdown();
         stream.unsubscribe();
-        CumulativeCommandEventCounterStream.reset();
+        RollingCommandEventCounterStream.reset();
     }
 
     @Test
     public void testEmptyStreamProducesZeros() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-A");
-        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-A");
+        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
@@ -73,20 +74,21 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
         } catch (InterruptedException ex) {
             fail("Interrupted ex");
         }
+        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertEquals(HystrixEventType.values().length, stream.getLatest().length);
         assertFalse(hasData(stream.getLatest()));
     }
 
     @Test
     public void testSingleSuccess() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-B");
-        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-B");
+        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
 
-        Command cmd = Command.from(groupKey, key, HystrixEventType.SUCCESS, 20);
+        CommandStreamTest.Command cmd = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 20);
 
         cmd.observe();
 
@@ -98,19 +100,20 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
         assertEquals(HystrixEventType.values().length, stream.getLatest().length);
         long[] expected = new long[HystrixEventType.values().length];
         expected[HystrixEventType.SUCCESS.ordinal()] = 1;
+        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testSingleFailure() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-C");
-        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-C");
+        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
 
-        Command cmd = Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
+        CommandStreamTest.Command cmd = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
 
         cmd.observe();
 
@@ -123,19 +126,20 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
         long[] expected = new long[HystrixEventType.values().length];
         expected[HystrixEventType.FAILURE.ordinal()] = 1;
         expected[HystrixEventType.FALLBACK_SUCCESS.ordinal()] = 1;
+        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testSingleTimeout() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-D");
-        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-D");
+        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
 
-        Command cmd = Command.from(groupKey, key, HystrixEventType.TIMEOUT);
+        CommandStreamTest.Command cmd = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.TIMEOUT);
 
         cmd.observe();
 
@@ -148,19 +152,20 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
         long[] expected = new long[HystrixEventType.values().length];
         expected[HystrixEventType.TIMEOUT.ordinal()] = 1;
         expected[HystrixEventType.FALLBACK_SUCCESS.ordinal()] = 1;
+        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testSingleBadRequest() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-E");
-        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-E");
+        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
 
-        Command cmd = Command.from(groupKey, key, HystrixEventType.BAD_REQUEST);
+        CommandStreamTest.Command cmd = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.BAD_REQUEST);
 
         cmd.observe();
 
@@ -173,21 +178,22 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
         long[] expected = new long[HystrixEventType.values().length];
         expected[HystrixEventType.BAD_REQUEST.ordinal()] = 1;
         expected[HystrixEventType.EXCEPTION_THROWN.ordinal()] = 1;
+        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testRequestFromCache() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-F");
-        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-F");
+        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
 
-        Command cmd1 = Command.from(groupKey, key, HystrixEventType.SUCCESS, 20);
-        Command cmd2 = Command.from(groupKey, key, HystrixEventType.RESPONSE_FROM_CACHE);
-        Command cmd3 = Command.from(groupKey, key, HystrixEventType.RESPONSE_FROM_CACHE);
+        CommandStreamTest.Command cmd1 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 20);
+        CommandStreamTest.Command cmd2 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.RESPONSE_FROM_CACHE);
+        CommandStreamTest.Command cmd3 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.RESPONSE_FROM_CACHE);
 
         cmd1.observe();
         cmd2.observe();
@@ -204,13 +210,14 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
         long[] expected = new long[HystrixEventType.values().length];
         expected[HystrixEventType.SUCCESS.ordinal()] = 1;
         expected[HystrixEventType.RESPONSE_FROM_CACHE.ordinal()] = 2;
+        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testShortCircuited() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-G");
-        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-G");
+        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
@@ -219,12 +226,12 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
         //3 failures in a row will trip circuit.  let bucket roll once then submit 2 requests.
         //should see 3 FAILUREs and 2 SHORT_CIRCUITs and then 5 FALLBACK_SUCCESSes
 
-        Command failure1 = Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
-        Command failure2 = Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
-        Command failure3 = Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
+        CommandStreamTest.Command failure1 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
+        CommandStreamTest.Command failure2 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
+        CommandStreamTest.Command failure3 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20);
 
-        Command shortCircuit1 = Command.from(groupKey, key, HystrixEventType.SUCCESS);
-        Command shortCircuit2 = Command.from(groupKey, key, HystrixEventType.SUCCESS);
+        CommandStreamTest.Command shortCircuit1 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS);
+        CommandStreamTest.Command shortCircuit2 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS);
 
         failure1.observe();
         failure2.observe();
@@ -253,13 +260,14 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
         expected[HystrixEventType.FAILURE.ordinal()] = 3;
         expected[HystrixEventType.SHORT_CIRCUITED.ordinal()] = 2;
         expected[HystrixEventType.FALLBACK_SUCCESS.ordinal()] = 5;
+        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testSemaphoreRejected() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-H");
-        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-H");
+        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
@@ -269,16 +277,16 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
         //submit 2 more requests and they should be SEMAPHORE_REJECTED
         //should see 10 SUCCESSes, 2 SEMAPHORE_REJECTED and 2 FALLBACK_SUCCESSes
 
-        List<Command> saturators = new ArrayList<Command>();
+        List<CommandStreamTest.Command> saturators = new ArrayList<CommandStreamTest.Command>();
 
         for (int i = 0; i < 10; i++) {
-            saturators.add(Command.from(groupKey, key, HystrixEventType.SUCCESS, 200, HystrixCommandProperties.ExecutionIsolationStrategy.SEMAPHORE));
+            saturators.add(CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 200, HystrixCommandProperties.ExecutionIsolationStrategy.SEMAPHORE));
         }
 
-        Command rejected1 = Command.from(groupKey, key, HystrixEventType.SUCCESS, 0, HystrixCommandProperties.ExecutionIsolationStrategy.SEMAPHORE);
-        Command rejected2 = Command.from(groupKey, key, HystrixEventType.SUCCESS, 0, HystrixCommandProperties.ExecutionIsolationStrategy.SEMAPHORE);
+        CommandStreamTest.Command rejected1 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 0, HystrixCommandProperties.ExecutionIsolationStrategy.SEMAPHORE);
+        CommandStreamTest.Command rejected2 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 0, HystrixCommandProperties.ExecutionIsolationStrategy.SEMAPHORE);
 
-        for (final Command saturator : saturators) {
+        for (final CommandStreamTest.Command saturator : saturators) {
             new Thread(new HystrixContextRunnable(new Runnable() {
                 @Override
                 public void run() {
@@ -310,13 +318,14 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
         expected[HystrixEventType.SUCCESS.ordinal()] = 10;
         expected[HystrixEventType.SEMAPHORE_REJECTED.ordinal()] = 2;
         expected[HystrixEventType.FALLBACK_SUCCESS.ordinal()] = 2;
+        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testThreadPoolRejected() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-I");
-        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-I");
+        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
@@ -326,16 +335,16 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
         //submit 2 more requests and they should be THREADPOOL_REJECTED
         //should see 10 SUCCESSes, 2 THREADPOOL_REJECTED and 2 FALLBACK_SUCCESSes
 
-        List<Command> saturators = new ArrayList<Command>();
+        List<CommandStreamTest.Command> saturators = new ArrayList<CommandStreamTest.Command>();
 
         for (int i = 0; i < 10; i++) {
-            saturators.add(Command.from(groupKey, key, HystrixEventType.SUCCESS, 200));
+            saturators.add(CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 200));
         }
 
-        Command rejected1 = Command.from(groupKey, key, HystrixEventType.SUCCESS, 0);
-        Command rejected2 = Command.from(groupKey, key, HystrixEventType.SUCCESS, 0);
+        CommandStreamTest.Command rejected1 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 0);
+        CommandStreamTest.Command rejected2 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 0);
 
-        for (final Command saturator : saturators) {
+        for (final CommandStreamTest.Command saturator : saturators) {
             saturator.observe();
         }
 
@@ -367,14 +376,14 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
 
     @Test
     public void testFallbackFailure() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-J");
-        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-J");
+        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
 
-        Command cmd = Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_FAILURE);
+        CommandStreamTest.Command cmd = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_FAILURE);
 
         cmd.observe();
 
@@ -388,19 +397,20 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
         expected[HystrixEventType.FAILURE.ordinal()] = 1;
         expected[HystrixEventType.FALLBACK_FAILURE.ordinal()] = 1;
         expected[HystrixEventType.EXCEPTION_THROWN.ordinal()] = 1;
+        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testFallbackMissing() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-K");
-        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-K");
+        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(10).subscribe(getSubscriber(latch));
 
-        Command cmd = Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_MISSING);
+        CommandStreamTest.Command cmd = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_MISSING);
 
         cmd.observe();
 
@@ -419,8 +429,8 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
 
     @Test
     public void testFallbackRejection() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-L");
-        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-L");
+        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
@@ -429,15 +439,15 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
         //fallback semaphore size is 5.  So let 5 commands saturate that semaphore, then
         //let 2 more commands go to fallback.  they should get rejected by the fallback-semaphore
 
-        List<Command> fallbackSaturators = new ArrayList<Command>();
+        List<CommandStreamTest.Command> fallbackSaturators = new ArrayList<CommandStreamTest.Command>();
         for (int i = 0; i < 5; i++) {
-            fallbackSaturators.add(Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_SUCCESS, 400));
+            fallbackSaturators.add(CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_SUCCESS, 400));
         }
 
-        Command rejection1 = Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_SUCCESS, 0);
-        Command rejection2 = Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_SUCCESS, 0);
+        CommandStreamTest.Command rejection1 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_SUCCESS, 0);
+        CommandStreamTest.Command rejection2 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 20, HystrixEventType.FALLBACK_SUCCESS, 0);
 
-        for (Command saturator: fallbackSaturators) {
+        for (CommandStreamTest.Command saturator: fallbackSaturators) {
             saturator.observe();
         }
 
@@ -461,13 +471,14 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
         expected[HystrixEventType.FALLBACK_SUCCESS.ordinal()] = 5;
         expected[HystrixEventType.FALLBACK_REJECTION.ordinal()] = 2;
         expected[HystrixEventType.EXCEPTION_THROWN.ordinal()] = 2;
+        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 
     @Test
     public void testCollapsed() {
         HystrixCommandKey key = HystrixCommandKey.Factory.asKey("BatchCommand");
-        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
+        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
         final CountDownLatch latch = new CountDownLatch(1);
@@ -491,18 +502,18 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
     }
 
     @Test
-    public void testMultipleEventsOverTimeGetStoredAndNeverAgeOut() {
-        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-CumulativeCounter-M");
-        stream = CumulativeCommandEventCounterStream.getInstance(key, 10, 100);
+    public void testMultipleEventsOverTimeGetStoredAndAgeOut() {
+        HystrixCommandKey key = HystrixCommandKey.Factory.asKey("CMD-RollingCounter-M");
+        stream = RollingCommandEventCounterStream.getInstance(key, 10, 100);
         stream.startCachingStreamValuesIfUnstarted();
 
-        //by doing a take(30), we ensure that no rolling out of window takes place
+        //by doing a take(30), we ensure that all rolling counts go back to 0
 
         final CountDownLatch latch = new CountDownLatch(1);
         stream.observe().take(30).subscribe(getSubscriber(latch));
 
-        Command cmd1 = Command.from(groupKey, key, HystrixEventType.SUCCESS, 20);
-        Command cmd2 = Command.from(groupKey, key, HystrixEventType.FAILURE, 10);
+        CommandStreamTest.Command cmd1 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.SUCCESS, 20);
+        CommandStreamTest.Command cmd2 = CommandStreamTest.Command.from(groupKey, key, HystrixEventType.FAILURE, 10);
 
         cmd1.observe();
         cmd2.observe();
@@ -514,9 +525,7 @@ public class CumulativeCommandEventCounterStreamTest extends CommandStreamTest {
         }
         assertEquals(HystrixEventType.values().length, stream.getLatest().length);
         long[] expected = new long[HystrixEventType.values().length];
-        expected[HystrixEventType.SUCCESS.ordinal()] = 1;
-        expected[HystrixEventType.FAILURE.ordinal()] = 1;
-        expected[HystrixEventType.FALLBACK_SUCCESS.ordinal()] = 1;
+        System.out.println("ReqLog : " + HystrixRequestLog.getCurrentRequest().getExecutedCommandsAsString());
         assertArrayEquals(expected, stream.getLatest());
     }
 }
